Fix 'Wconversion' warns: static casting data types for mwcc monitoredqueue and orderedhashmap

+++ src/groups/mwc/mwcc/mwcc_monitoredqueue.h
@@ -590,7 +590,7 @@ template <class QUEUE, class QUEUE_TRAITS>
 inline int MonitoredQueue<QUEUE, QUEUE_TRAITS>::pushBack(
     bslmf::MovableRef<ElementType> value)
 {
-    const int currentQueueLen = numElements();
+    const int currentQueueLen = static_cast<int>(numElements());
     if (currentQueueLen + 1 >= capacity()) {
         // We've filled the queue.  Alarm
         if (!Traits::isPushBackDisabled(d_queue) &&
